Fix formato de numero

+++ app/Http/Controllers/HomeController.php
@@ -41,7 +41,7 @@ class HomeController extends Controller
         $detalle = App\Game::findOrFail($id);
         $comments = DB::table('comments')->where('comments.game_id', $id)
                                          ->get();
-        $price_rent = ((int)($detalle->amount * $detalle->percentaje_rent) / 100);
+        $price_rent = ($detalle->amount * $detalle->percentaje_rent) / 100;
 		return view('alquilar', compact('detalle', 'comments', 'price_rent'));
     }
